Seperating user list with overview
