+++ angular/projects/admin/src/app/admin/user/user-list/user-list.component.html
@@ -1,15 +1,14 @@
-<mat-tab-group>
-  <mat-tab label="Users">
-    <tanam-user-list></tanam-user-list>
-  </mat-tab>
-  <mat-tab label="Invited">
-    <mat-card>
-      <mat-list dense>
-        <mat-list-item>
+<mat-card>
+    <mat-list dense>
+      <virtual-scroller #scroll
+                        [items]="items"
+                        (vsEnd)="fetchMore($event)"
+                        [enableUnequalChildrenSizes]="true">
+        <mat-list-item *ngFor="let item of scroll.viewPortItems">
           <img matListAvatar
-               src="https://picsum.photos/200">
-          <h4 mat-line>Putu Deta</h4>
-          <p mat-line>deta@oddbit.id</p>
+               src="{{item.photoUrl || 'assets/icons/tanam-icon.svg'}}">
+          <h4 mat-line>{{item.name}}</h4>
+          <p mat-line>{{item.email}}</p>
           <mat-form-field>
             <mat-label>Role</mat-label>
             <mat-select value="admin">
@@ -24,7 +23,6 @@
           </button>
           <mat-divider></mat-divider>
         </mat-list-item>
-      </mat-list>
-    </mat-card>
-  </mat-tab>
-</mat-tab-group>
+      </virtual-scroller>
+    </mat-list>
+  </mat-card>
