+++ angular/projects/admin/src/app/admin/user/user.module.ts
@@ -2,10 +2,12 @@ import { NgModule } from '@angular/core';
 import { UserOverviewComponent } from './user-overview/user-overview.component';
 import { AppMaterialModule } from '../../app-material.module';
 import { VirtualScrollerModule } from 'ngx-virtual-scroller';
+import { UserListComponent } from './user-list/user-list.component';
 
 @NgModule({
   declarations: [
-    UserOverviewComponent
+    UserOverviewComponent,
+    UserListComponent
   ],
   imports: [
     AppMaterialModule,
